backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
azure-search-documents==11.4.0
azure-core==1.30.0
openai==1.54.0
httpx==0.27.0
azure-ai-documentintelligence==1.0.0b1
azure-storage-blob==12.19.0
python-multipart==0.0.6
redis==5.0.1
slowapi==0.1.9
tenacity==8.2.3
gunicorn==21.2.0
httpx[http2]
